requirements.txt
# --- Web UI ---
streamlit==1.36.0

# --- Data ---
pandas==2.2.3
numpy==2.0.2
matplotlib==3.9.0
sqlalchemy==2.0.30
pydantic==2.7.1

# --- HTTP & Parsing ---
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.47.0

# --- Notifications ---
python-telegram-bot==20.7

# --- Machine Learning (future use) ---
scikit-learn==1.5.1
lightgbm==4.5.0

# --- Utils ---
pyyaml==6.0.2
Unidecode==1.3.8
